requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-phonenumber-field==7.0.2
phonenumbers==8.12.47
Pillow==9.4.0
python-dotenv==0.21.1
django-cors-headers==3.13.0
django-filter==22.1
django-ckeditor==6.5.1

django-extensions==3.2.1
pydot==1.4.2
pyparsing==3.0.9
pre-commit==3.1.1
autopep8==2.0.1
black==23.1.0
flake8==6.0.0
